refactor: restructure watering components for clarity
