+++ frontend/src/components/trees/add-watering/WaterTreeButton.svelte
