+++ frontend/src/components/trees/add-watering/WateringSuccess.svelte
