+++ frontend/src/components/trees/index.ts
@@ -1,7 +1,7 @@
 export { default as AdoptTreeButton } from './AdoptTreeButton.svelte';
 export { default as TreeMetric } from './TreeMetric.svelte';
+export { default as TreeWaterings } from './TreeWaterings.svelte';
 export { default as WaterColumn } from './WaterColumn.svelte';
-export { default as WaterTreeButton } from './watering/WaterTreeButton.svelte';
-export { default as TreeWaterings } from './watering/TreeWaterings.svelte';
-export { default as AddWateringForm} from './watering/AddWateringForm.svelte';
-export { default as WateringHistory } from './watering/WateringHistory.svelte';
+export { default as AddWateringForm } from './add-watering/AddWateringForm.svelte';
+export { default as WateringSuccess } from './add-watering/WateringSuccess.svelte';
+export { default as WaterTreeButton } from './add-watering/WaterTreeButton.svelte';
