+++ frontend/src/components/waterings/ConfirmDeleteWatering.svelte
