+++ frontend/src/components/waterings/DeleteWateringButton.svelte
