+++ frontend/src/components/waterings/WateringCard.svelte
