+++ frontend/src/components/waterings/WateringHistory.svelte
