+++ frontend/src/components/waterings/WateringTable.svelte
