+++ frontend/src/components/waterings/index.ts
@@ -1,0 +1,1 @@
+export { default as WateringHistory } from './WateringHistory.svelte';
